read unsigned int 32 for length fix issue with skipping bytes in deeper EventIO levels

+++ src/main/java/streams/cta/io/Event/FullEvent.java
@@ -78,7 +78,8 @@ public class FullEvent {
         //TODO should we use the header to skip this whole event item in a case any of its subitems are failed to be read?
         if (header.getVersion() != 0) {
             log.error("Unsupported FullEvent version: " + header.getVersion());
-            buffer.skipBytes(header.getLength());
+            //TODO we should get item end?!
+            buffer.skipBytes((int) header.getLength());
             return null;
         }
